ai-assistants/pyproject.toml
[project]
name = "ai-assistants"
version = "0.1.0"
description = "AI Underwriter Assistants"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "catboost>=1.2.7",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "openai>=1.66.5",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "scikit-learn>=1.6.1",
]

[dependency-groups]
dev = [
    "pydocstyle>=6.3.0",
    "ruff>=0.11.0",
]
